Now using openghg_calscales for scale conversion, fixes Issue #6
