--- data/scale_convert.csv
@@ -1,8 +1,0 @@
-﻿Species,SIO-98/SIO-93,SIO-05/SIO-98,SIO-16/SIO-05,TU-87/CSIRO-94,WMO-X2004A/TU-87,WMO-X2004A/NOAA-X2004A
-cfc-11,1.0082,0.9945,,,,
-cfc-12,1.0053,1,,,,
-cfc-113,1.0038,1.0042,,,,
-ch3ccl3,1.0184,0.9957,,,,
-ccl4,1.009,1,,,,
-ch4,,,,1.0119,1.002,1
-n2o,1.0058,1,1,,,
