+++ requirements.txt
@@ -11,7 +11,6 @@ nbformat
 h5netcdf
 dvc
 dvc-gdrive
-networkx
 
 # for testing
 cfchecker
